move evaluator to separate class
